Mini Update for most of classes
Audio: Created getters and setters for the name and path. Added a copy
constructor.

Component: refactored variable names for a cleaner convention.

Console: cleaned up the useless sprintToConsole static function, wasnt
using it, didnt implement it correctly , so goodbye.

Debugger:  Created a legitimate class with a public QDebug, QString and
console. Im hoping to redirect all qDebug errors to the console.

Engine: Added the debugger

Entity: Got rid of ComponentMessageHandler failed attempt of creating
Messaging between Components.

Keyboard: refactored coding conventions

Screen: created a wrapper to the version of OpenGL Used on the screen.
This is set into a vector 2 of integers.

Serializable: an abstract object which can be serialized into a json
Document to be saved somewhere

Vector2: changed the ToString returning a constant character pointer to
a basic std::string